Correct -Wformat-pedantic warnings
Many of the warnings were newly visible after the DBPRINT macro
was modified, but the pedantic warnings also caught existing printf
family usage which was not strictly correct.

Mostly replacing 0x%x with %p for pointers, adding (void *) for
those and other pointers, %td for ptrdiff_t, but also changing
some function parameters from UNSIGNED (uintptr_t) to int
where that was the more appropriate type.

+++ src/timer.c
@@ -473,7 +473,7 @@ static void int_timer_init()
   timert.it_interval.tv_usec = timert.it_value.tv_usec = TIMER_INTERVAL;
   setitimer(ITIMER_VIRTUAL, &timert, NULL);
 
-  DBPRINT(("Timer interval set to %d usec\n", timert.it_value.tv_usec));
+  DBPRINT(("Timer interval set to %ld usec\n", (long)timert.it_value.tv_usec));
 #endif /* DOS */
 }
